requirements.txt
robotframework-seleniumlibrary
robotframework-jsonlibrary
robotframework-sikulilibrary
webdriver-manager
python-dotenv